+++ pyproject.toml
@@ -16,6 +16,7 @@ uvicorn = "^0.32.0"
 python-dotenv = "^1.0.1"
 pydantic-settings = "^2.6.1"
 alembic = "^1.14.0"
+greenlet = "^3.1.1"
 
 
 [build-system]
