+++ src/models/__init__.py
@@ -1,0 +1,14 @@
+from .base import Base
+from .user import User
+from .coords import Coords
+from .pereval import PerevalAdded, Status
+from .images import PerevalImages
+
+__all__ = [
+    "Base",
+    "User",
+    "Coords",
+    "PerevalAdded",
+    "PerevalImages",
+    "Status"
+]
