+++ src/models/images.py
@@ -1,0 +1,13 @@
+from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy.orm import relationship
+from .base import Base
+
+
+class PerevalImages(Base):
+    __tablename__ = "pereval_images"
+
+    id = Column(Integer, primary_key=True)
+    pereval_id = Column(Integer, ForeignKey("pereval_added.id"), nullable=False)
+    image_url = Column(String, nullable=False)
+
+    pereval = relationship("PerevalAdded", back_populates="images")
